backend/requirements.txt
Django==4.2.23
django-cors-headers==4.3.1
openai>=1.0.0
python-dotenv==1.0.0
